+++ scripts/solr-cloud-entrypoint.sh
@@ -1,13 +1,9 @@
 #!/bin/bash
-# Copyright (c) 2026 Eledia GmbH / Bernd Schreistetter
-# SPDX-License-Identifier: MIT
+# Copyright (c) 2026 eLeDia.de / Bernd Schreistetter (bsc)
 # Version: v3.0.1
 
 # =========================================
 # Solr runtime entrypoint
-# Developer: BSC Bernd Schreistetter
-# Company: Eledia.de
-# Version: v3.0.1
 # =========================================
 # Standalone mode:
 #   exec solr-foreground directly.